processor/pyproject.toml
[tool.poetry]
name = "pycontainer"
version = "0.1.0"
description = ""
authors = ["Keith Nordstrom <keith@timeli.io>"]
readme = "README.rst"
packages = [{include = "library"}, {include = "processor"}]

[tool.poetry.scripts]
processor = "processor.main:start"
event_processor = "processor.main:start_kafka_calendar"

[tool.poetry.dependencies]
py2neo = "*"
python = "^3.11"
pytest = "^8.1.1"
Flask = "^3.0.2"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
weaviate-client = "^4.5.4"
langchain = "^0.1.12"
langchain-experimental = "^0.0.55"
langchain-openai = "^0.0.8"
beautifulsoup4 = "^4.12.3"
email-reply-parser = "^0.5.12"
coverage = "^7.4.4"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.1"
ics = "^0.7.2"




[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
